+++ CotEditor/Sources/Text Finder/Multiple Replace/Views/MultipleReplaceListView.swift
@@ -51,12 +51,11 @@ struct MultipleReplaceListView: View {
             ForEach(self.settingNames, id: \.self) { name in
                 SettingNameField(text: name) { newName in
                     do {
-                        try self.manager.renameSetting(name: name, to: newName)
+                        self.selection = try self.manager.renameSetting(name: name, to: newName)
                     } catch {
                         self.error = error
                         return false
                     }
-                    self.selection = newName
                     return true
                 }
                 .focused($editingItem, equals: name)